Add chunks library.

+++ src/action/finder/action.rs
@@ -91,7 +91,7 @@ impl ChunkId {
 #[derive(Debug)]
 struct DownloadEntry {
     chunk_ids: Vec<ChunkId>,
-    chunk_size: u64,
+    chunk_size: usize,
 }
 
 impl From<DownloadEntry> for rrg_proto::BlobImageDescriptor {
@@ -106,7 +106,7 @@ impl From<DownloadEntry> for rrg_proto::BlobImageDescriptor {
                     digest: Some(x.sha256.to_vec()),
                 })
                 .collect::<Vec<_>>(),
-            chunk_size: Some(e.chunk_size),
+            chunk_size: Some(e.chunk_size as u64),
         }
     }
 }
